+++ upload_run.py
@@ -15,7 +15,7 @@ argument_parser = argparse.ArgumentParser(
     formatter_class=argparse.RawDescriptionHelpFormatter,
     description=textwrap.dedent('''
     This program parses sequencing runs and uploads them to IRIDA.
-    
+
     required arguments:
       --d DIRECTORY, --directory DIRECTORY
                             Location of sequencing run to upload.
@@ -28,7 +28,7 @@ argument_parser = argparse.ArgumentParser(
 # a optional argument, with the required set to True. We have to suppress the help on the argument, and add it's help
 # information to the formatted description text of the parser for the argument to be shown as required when --help
 # is used.
-argument_parser.add_argument('-d','--directory',
+argument_parser.add_argument('-d', '--directory',
                              action='store',
                              required=True,
                              help=argparse.SUPPRESS)