+++ CHANGELOG.md
@@ -7,10 +7,14 @@ Added functionality:
 * Added config option overrides for the command line
     * Can define in command line `--option parameter` or without a parameter `--option` for a user prompt
     * Use the `--help` option for more details on these options
+* The `directory` parser can now handle full (absolute) file paths in the sample list file
 
 Bug Fixes:
 * Fixed hard crash that sometimes occurred when valid url's were given as the base_url
 * Fixed hard crash when running the miniseq parser on a miseq run directory
+
+API Changes:
+* Added 400, 401, 403, and 500 to http errors we explicitly have exceptions for. (Other errors still generate generic exceptions that are handled the same way.)
 
 Beta 0.2.1
 ----------